The Ultimate Photo Resistor Pain In The Ass Block Detector Version 2.0!
I calibrate every time you run the code. USAGE: On compiling, you have 30 seconds to put the white block in the sensor. The sensor collects that in another 30 seconds. We also do this same thing with the blue block afterwords. I could map it to the button on the arduino (instead of 30 seconds and relying on time). After that, it will sort every single value in the array by using a smiliarity formula which outputs a percentage of, not the color, but the difference in color. So, whichever percentage is closer to 0 corresponds to the color it represents.